+++ src/components/StatsExportModal.tsx
@@ -1,3 +1,4 @@
+import { getAlbumList } from '../api/subsonicLibrary';
 import type { SubsonicAlbum } from '../api/subsonicTypes';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
@@ -5,7 +6,6 @@ import { useTranslation } from 'react-i18next';
 import { X } from 'lucide-react';
 import { save } from '@tauri-apps/plugin-dialog';
 import { writeFile } from '@tauri-apps/plugin-fs';
-import { getAlbumList } from '../api/subsonic';
 import { showToast } from '../utils/toast';
 import {
   exportAlbumCardBlob,
