+++ src/hooks/useOrbitHost.ts
@@ -1,8 +1,8 @@
+import { getSong } from '../api/subsonicLibrary';
 import { songToTrack } from '../utils/songToTrack';
 import { useEffect, useRef } from 'react';
 import { useOrbitStore } from '../store/orbitStore';
 import { usePlayerStore } from '../store/playerStore';
-import { getSong } from '../api/subsonic';
 import {
   writeOrbitState,
   writeOrbitHeartbeat,
