+++ src/pages/ComposerDetail.tsx
@@ -1,7 +1,8 @@
+import { getArtist, getArtistInfo } from '../api/subsonicArtists';
 import type { SubsonicArtist, SubsonicAlbum, SubsonicArtistInfo } from '../api/subsonicTypes';
 import { useEffect, useState, useMemo } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
-import { getArtist, getArtistInfo, star, unstar, buildCoverArtUrl, coverArtCacheKey } from '../api/subsonic';
+import { star, unstar, buildCoverArtUrl, coverArtCacheKey } from '../api/subsonic';
 import { ndListAlbumsByArtistRole } from '../api/navidromeBrowse';
 import AlbumCard from '../components/AlbumCard';
 import CachedImage from '../components/CachedImage';
