+++ src/utils/buildInfiniteQueueCandidates.test.ts
@@ -6,12 +6,17 @@
  * refactor (2026-05-12). This test pins the artist-first / random-fallback
  * order, the dedup contract against existingIds, and the autoAdded flag.
  */
+import { getSimilarSongs2, getTopSongs } from '../api/subsonicArtists';
+import { getRandomSongs } from '../api/subsonicLibrary';
 import type { Track } from '../store/playerStoreTypes';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-vi.mock('../api/subsonic', () => ({
+vi.mock('../api/subsonicArtists', () => ({
   getSimilarSongs2: vi.fn(),
   getTopSongs: vi.fn(),
+}));
+
+vi.mock('../api/subsonicLibrary', () => ({
   getRandomSongs: vi.fn(),
 }));
 
@@ -22,7 +27,6 @@ vi.mock('./mixRatingFilter', () => ({
 }));
 
 import { buildInfiniteQueueCandidates } from './buildInfiniteQueueCandidates';
-import { getRandomSongs, getSimilarSongs2, getTopSongs } from '../api/subsonic';
 import {
   enrichSongsForMixRatingFilter,
   getMixMinRatingsConfigFromAuth,
